Fix generated files
