+++ cmd/piper_test.go
@@ -2,12 +2,10 @@ package cmd
 
 import (
 	"fmt"
-	"io"
+	"github.com/SAP/jenkins-library/pkg/mock"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 	"testing"
 
 	"github.com/SAP/jenkins-library/pkg/config"
@@ -15,157 +13,6 @@ import (
 	flag "github.com/spf13/pflag"
 	"github.com/stretchr/testify/assert"
 )
-
-type execMockRunner struct {
-	dir                 []string
-	env                 [][]string
-	calls               []execCall
-	stdout              io.Writer
-	stderr              io.Writer
-	stdoutReturn        map[string]string
-	shouldFailOnCommand map[string]error
-}
-
-type execCall struct {
-	exec   string
-	params []string
-}
-
-type shellMockRunner struct {
-	dir                 string
-	env                 [][]string
-	calls               []string
-	shell               []string
-	stdout              io.Writer
-	stderr              io.Writer
-	stdoutReturn        map[string]string
-	shouldFailOnCommand map[string]error
-}
-
-func (m *execMockRunner) Dir(d string) {
-	m.dir = append(m.dir, d)
-}
-
-func (m *execMockRunner) Env(e []string) {
-	m.env = append(m.env, e)
-}
-
-func (m *execMockRunner) RunExecutable(e string, p ...string) error {
-
-	exec := execCall{exec: e, params: p}
-	m.calls = append(m.calls, exec)
-
-	c := strings.Join(append([]string{e}, p...), " ")
-
-	return handleCall(c, m.stdoutReturn, m.shouldFailOnCommand, m.stdout)
-}
-
-func (m *execMockRunner) Stdout(out io.Writer) {
-	m.stdout = out
-}
-
-func (m *execMockRunner) Stderr(err io.Writer) {
-	m.stderr = err
-}
-
-func (m *shellMockRunner) Dir(d string) {
-	m.dir = d
-}
-
-func (m *shellMockRunner) Env(e []string) {
-	m.env = append(m.env, e)
-}
-
-func (m *shellMockRunner) RunShell(s string, c string) error {
-
-	m.shell = append(m.shell, s)
-	m.calls = append(m.calls, c)
-
-	return handleCall(c, m.stdoutReturn, m.shouldFailOnCommand, m.stdout)
-}
-
-func handleCall(call string, stdoutReturn map[string]string, shouldFailOnCommand map[string]error, stdout io.Writer) error {
-
-	if stdoutReturn != nil {
-		for k, v := range stdoutReturn {
-
-			found := k == call
-
-			if !found {
-
-				r, e := regexp.Compile(k)
-				if e != nil {
-					return e
-					// we don't distinguish here between an error returned
-					// since it was configured or returning this error here
-					// indicating an invalid regex. Anyway: when running the
-					// test we will see it ...
-				}
-				if r.MatchString(call) {
-					found = true
-
-				}
-			}
-
-			if found {
-				stdout.Write([]byte(v))
-			}
-		}
-	}
-
-	if shouldFailOnCommand != nil {
-		for k, v := range shouldFailOnCommand {
-
-			found := k == call
-
-			if !found {
-				r, e := regexp.Compile(k)
-				if e != nil {
-					return e
-					// we don't distinguish here between an error returned
-					// since it was configured or returning this error here
-					// indicating an invalid regex. Anyway: when running the
-					// test we will see it ...
-				}
-				if r.MatchString(call) {
-					found = true
-
-				}
-			}
-
-			if found {
-				return v
-			}
-		}
-	}
-
-	return nil
-}
-
-func (m *shellMockRunner) Stdout(out io.Writer) {
-	m.stdout = out
-}
-
-func (m *shellMockRunner) Stderr(err io.Writer) {
-	m.stderr = err
-}
-
-type stepOptions struct {
-	TestParam string `json:"testParam,omitempty"`
-}
-
-func openFileMock(name string) (io.ReadCloser, error) {
-	var r string
-	switch name {
-	case "testDefaults.yml":
-		r = "general:\n  testParam: testValue"
-	case "testDefaultsInvalid.yml":
-		r = "invalid yaml"
-	default:
-		r = ""
-	}
-	return ioutil.NopCloser(strings.NewReader(r)), nil
-}
 
 func TestAddRootFlags(t *testing.T) {
 	var testRootCmd = &cobra.Command{Use: "test", Short: "This is just a test"}
@@ -189,7 +36,7 @@ func TestPrepareConfig(t *testing.T) {
 		stepConfigJSONBak := GeneralConfig.StepConfigJSON
 		GeneralConfig.StepConfigJSON = `{"testParam": "testValueJSON"}`
 		defer func() { GeneralConfig.StepConfigJSON = stepConfigJSONBak }()
-		testOptions := stepOptions{}
+		testOptions := mock.StepOptions{}
 		var testCmd = &cobra.Command{Use: "test", Short: "This is just a test"}
 		testCmd.Flags().StringVar(&testOptions.TestParam, "testParam", "", "test usage")
 		metadata := config.StepData{
@@ -202,13 +49,13 @@ func TestPrepareConfig(t *testing.T) {
 			},
 		}
 
-		PrepareConfig(testCmd, &metadata, "testStep", &testOptions, openFileMock)
+		PrepareConfig(testCmd, &metadata, "testStep", &testOptions, mock.OpenFileMock)
 		assert.Equal(t, "testValueJSON", testOptions.TestParam, "wrong value retrieved from config")
 	})
 
 	t.Run("using config files", func(t *testing.T) {
 		t.Run("success case", func(t *testing.T) {
-			testOptions := stepOptions{}
+			testOptions := mock.StepOptions{}
 			var testCmd = &cobra.Command{Use: "test", Short: "This is just a test"}
 			testCmd.Flags().StringVar(&testOptions.TestParam, "testParam", "", "test usage")
 			metadata := config.StepData{
@@ -221,7 +68,7 @@ func TestPrepareConfig(t *testing.T) {
 				},
 			}
 
-			err := PrepareConfig(testCmd, &metadata, "testStep", &testOptions, openFileMock)
+			err := PrepareConfig(testCmd, &metadata, "testStep", &testOptions, mock.OpenFileMock)
 			assert.NoError(t, err, "no error expected but error occured")
 
 			//assert config
@@ -237,11 +84,11 @@ func TestPrepareConfig(t *testing.T) {
 
 		t.Run("error case", func(t *testing.T) {
 			GeneralConfig.DefaultConfig = []string{"testDefaultsInvalid.yml"}
-			testOptions := stepOptions{}
+			testOptions := mock.StepOptions{}
 			var testCmd = &cobra.Command{Use: "test", Short: "This is just a test"}
 			metadata := config.StepData{}
 
-			err := PrepareConfig(testCmd, &metadata, "testStep", &testOptions, openFileMock)
+			err := PrepareConfig(testCmd, &metadata, "testStep", &testOptions, mock.OpenFileMock)
 			assert.Error(t, err, "error expected but none occured")
 		})
 	})
